Refactor/meta (#194)
* small hook improvements

* refactor programIds

* split contexts/meta to separate files

* process account right after they loaded
