+++ js/packages/web/src/contexts/meta/isMetadataPartOfStore.ts
@@ -1,0 +1,26 @@
+import { Metadata, ParsedAccount } from '@oyster/common';
+import { Store, WhitelistedCreator } from '../../models/metaplex';
+
+export const isMetadataPartOfStore = (
+  m: ParsedAccount<Metadata>,
+  store: ParsedAccount<Store> | null,
+  whitelistedCreatorsByCreator: Record<
+    string,
+    ParsedAccount<WhitelistedCreator>
+  >,
+  useAll: boolean,
+) => {
+  if (useAll) {
+    return true;
+  }
+  if (!m?.info?.data?.creators || !store?.info) {
+    return false;
+  }
+
+  return m.info.data.creators.some(
+    c =>
+      c.verified &&
+      (store.info.public ||
+        whitelistedCreatorsByCreator[c.address.toBase58()]?.info?.activated),
+  );
+};
