+++ js/packages/web/src/utils/pubkeyToString.ts
@@ -1,0 +1,5 @@
+import { PublicKey } from '@solana/web3.js';
+
+export const pubkeyToString = (key: PublicKey | string = '') => {
+  return typeof key === 'string' ? key : key?.toBase58() || '';
+};
